# api/openapi.yaml
openapi: 3.0.3
info:
  title: Invite API
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /dashboard/stats:
    get:
      summary: Get dashboard statistics and activity
      operationId: GetDashboardStats
      responses:
        '200':
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
        '401':
          description: Not authenticated
...
  /auth/login:
    post:
      summary: Log in
      operationId: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Logged in
        '401':
          description: Invalid credentials
...
  /auth/forgot-password:
    post:
      summary: Request password reset
      operationId: ForgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '204':
          description: Reset email sent
  /auth/reset-password:
    post:
      summary: Reset password
      operationId: ResetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
      responses:
        '204':
          description: Password reset successful
        '400':
          description: Invalid or expired token
...
components:
  schemas:
...
    DashboardStats:
      type: object
      required: [stats, bottlenecks, activity]
      properties:
        stats:
          type: object
          properties:
            active_invites: { type: integer }
            failed_emails: { type: integer }
            success_rate: { type: number, format: float }
        timeline:
          type: array
          items:
            type: object
            required: [id, title, status, phases]
            properties:
              id: { type: string, format: uuid }
              title: { type: string }
              status: { type: string }
              phases:
                type: array
                items:
                  type: object
                  required: [order, status, accepted_count, declined_count, total_invitees]
                  properties:
                    order: { type: integer }
                    status: { type: string }
                    accepted_count: { type: integer }
                    declined_count: { type: integer }
                    total_invitees: { type: integer }
        bottlenecks:
          type: array
          items:
            type: object
            properties:
              invite_id: { type: string, format: uuid }
              title: { type: string }
              phase_order: { type: integer }
              strategy_kind: { type: string }
              waiting_for: { type: string }
              active_since: { type: string, format: date-time }
              tags:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
        activity:
          type: array
          items:
            type: object
            properties:
              timestamp: { type: string, format: date-time }
              type: { type: string }
              message: { type: string }
    Tag:
      type: object
      required: [id, name, color]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        color: { type: string }
...